backend/dev-requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-db-logger==0.1.12
django-extensions==3.2.3
django-filter==23.2
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
model-bakery==1.12.0
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
Unidecode==1.3.6
